more cleanup

+++ GMap.NET/GMap.NET.Core/MapProviders/CzechOld/CzechHistoryMapProvider.cs
@@ -69,7 +69,7 @@ namespace GMap.NET.MapProviders
             // http://m4.mapserver.mapy.cz/army2/9_7d00000_8080000
 
             long xx = pos.X << (28 - zoom);
-            long yy = ((long)Math.Pow(2.0, (double)zoom) - 1 - pos.Y) << (28 - zoom);
+            long yy = ((long)Math.Pow(2.0, zoom) - 1 - pos.Y) << (28 - zoom);
 
             return string.Format(UrlFormat, GetServerNum(pos, 3) + 1, zoom, xx, yy);
         }
